bugfixes.patch from ragnarok2040, http://forums.ps2dev.org/viewtopic.php?t=12944

+++ ee/kernel/include/timer.h
@@ -19,25 +19,25 @@ extern "C" {
 #endif
 
 // EE Timers
-#define T0_COUNT   ((volatile unsigned long*)0x10000000) 
-#define T0_MODE      ((volatile unsigned long*)0x10000010) 
-#define T0_COMP      ((volatile unsigned long*)0x10000020) 
-#define T0_HOLD      ((volatile unsigned long*)0x10000030) 
+#define T0_COUNT   ((volatile unsigned int*)0x10000000) 
+#define T0_MODE      ((volatile unsigned int*)0x10000010) 
+#define T0_COMP      ((volatile unsigned int*)0x10000020) 
+#define T0_HOLD      ((volatile unsigned int*)0x10000030) 
 
-#define T1_COUNT   ((volatile unsigned long*)0x10000800) 
-#define T1_MODE      ((volatile unsigned long*)0x10000810) 
-#define T1_COMP      ((volatile unsigned long*)0x10000820) 
-#define T1_HOLD      ((volatile unsigned long*)0x10000830) 
+#define T1_COUNT   ((volatile unsigned int*)0x10000800) 
+#define T1_MODE      ((volatile unsigned int*)0x10000810) 
+#define T1_COMP      ((volatile unsigned int*)0x10000820) 
+#define T1_HOLD      ((volatile unsigned int*)0x10000830) 
 
 // Note! T2 and T3 don't have a Tn_HOLD register! 
 // ---------------------------------------------- 
-#define T2_COUNT   ((volatile unsigned long*)0x10001000) 
-#define T2_MODE      ((volatile unsigned long*)0x10001010) 
-#define T2_COMP      ((volatile unsigned long*)0x10001020) 
+#define T2_COUNT   ((volatile unsigned int*)0x10001000) 
+#define T2_MODE      ((volatile unsigned int*)0x10001010) 
+#define T2_COMP      ((volatile unsigned int*)0x10001020) 
 
-#define T3_COUNT   ((volatile unsigned long*)0x10001800) 
-#define T3_MODE      ((volatile unsigned long*)0x10001810) 
-#define T3_COMP      ((volatile unsigned long*)0x10001820) 
+#define T3_COUNT   ((volatile unsigned int*)0x10001800) 
+#define T3_MODE      ((volatile unsigned int*)0x10001810) 
+#define T3_COMP      ((volatile unsigned int*)0x10001820) 
 
 #define Tn_MODE(CLKS,GATE,GATS,GATM,ZRET,CUE,CMPE,OVFE,EQUF,OVFF) \
     (u32)((u32)(CLKS) | ((u32)(GATE) << 2) | \
